refactor(order): modularize order management controller
